+++ src/interfaces/MumbleFormValues.ts
@@ -1,0 +1,3 @@
+export type FormValues = {
+  text: string
+}
